update costing and types

+++ wheel/python/chia_rs/chia_rs.pyi
@@ -4468,8 +4468,8 @@ class ConsensusConstants:
     PLOT_STRENGTH_7_HEIGHT: uint32
     PLOT_STRENGTH_8_HEIGHT: uint32
     QUALITY_PROOF_SCAN_FILTER: uint8
-    SHATREE_RECURSE_COST: uint32
-    SHATREE_COST_PER_BYTE: uint32
+    SHATREE_RECURSE_COST: uint64
+    SHATREE_COST_PER_BYTE: uint64
     def __new__(
         cls,
         SLOT_BLOCKS_TARGET: uint32,
@@ -4528,8 +4528,8 @@ class ConsensusConstants:
         PLOT_STRENGTH_7_HEIGHT: uint32,
         PLOT_STRENGTH_8_HEIGHT: uint32,
         QUALITY_PROOF_SCAN_FILTER: uint8,
-        SHATREE_RECURSE_COST: uint32,
-        SHATREE_COST_PER_BYTE: uint32
+        SHATREE_RECURSE_COST: uint64,
+        SHATREE_COST_PER_BYTE: uint64
     ) -> ConsensusConstants: ...
     def __hash__(self) -> int: ...
     def __repr__(self) -> str: ...
@@ -4604,5 +4604,5 @@ class ConsensusConstants:
         PLOT_STRENGTH_7_HEIGHT: Union[ uint32, _Unspec] = _Unspec(),
         PLOT_STRENGTH_8_HEIGHT: Union[ uint32, _Unspec] = _Unspec(),
         QUALITY_PROOF_SCAN_FILTER: Union[ uint8, _Unspec] = _Unspec(),
-        SHATREE_RECURSE_COST: Union[ uint32, _Unspec] = _Unspec(),
-        SHATREE_COST_PER_BYTE: Union[ uint32, _Unspec] = _Unspec()) -> ConsensusConstants: ...
+        SHATREE_RECURSE_COST: Union[ uint64, _Unspec] = _Unspec(),
+        SHATREE_COST_PER_BYTE: Union[ uint64, _Unspec] = _Unspec()) -> ConsensusConstants: ...
